+++ Pricol2/HPP/Resources.h
@@ -9,7 +9,7 @@
 #include "array"
 
 const int TEXTURE_SIZE = 128, GUN_SIZE = 1, ICON_SIZE = 64, TEXTURE_COUNT = 15;
-constexpr float SCREEN_W = 1280.0f, SCREEN_H = 720.0f;
+constexpr int SCREEN_W = 1280, SCREEN_H = 720;
 constexpr int SPACE_SIZE_W = 40, SPACE_SIZE_H = 40;
 
 class Resources
